compose button code
